refactor(scene_management): use dedicated scene database operations
- Introduced SceneDatabaseOperations class for all scene DB interactions
- Moved scene DB logic from SceneManagementAPI to SceneDatabaseOperations
- Added scene_models module for data models like SceneIdentifier and results
- Added scene_validator module for Fountain format validation
- Updated SceneManagementAPI to delegate DB calls to SceneDatabaseOperations
- Removed redundant DB methods from SceneManagementAPI
- Updated imports and tests to use new models and DB operations

This refactor improves separation of concerns, code organization, and testability.
