+++ tests/test_scene_cli_extended.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock, patch
 
 from typer.testing import CliRunner
 
-from scriptrag.api.scene_management import (
+from scriptrag.api.scene_models import (
     AddSceneResult,
     BibleReadResult,
     DeleteSceneResult,
